+++ src/main/java/com/saisai/domain/course/service/CourseService.java
@@ -4,6 +4,7 @@ import static com.saisai.domain.common.exception.ExceptionCode.COURSE_NOT_FOUND;
 
 import com.saisai.domain.common.exception.CustomException;
 import com.saisai.domain.common.utils.s3.ImageUtil;
+import com.saisai.domain.common.aws.s3.ImageUtil;
 import com.saisai.domain.course.dto.projection.CoursePageProjection;
 import com.saisai.domain.course.dto.response.CourseDetailsRes;
 import com.saisai.domain.course.dto.response.CoursePageRes;
